fix accessibility for side nav and code cleanup
