+++ lib/components/SideNav/index.js
@@ -1,4 +1,4 @@
-import React, { useRef, useState, useEffect, createRef } from "react";
+import React, { useRef, useState, useEffect } from "react";
 import styled from "styled-components";
 import { css } from "@styled-system/css";
 import { themeGet } from "@styled-system/theme-get";
@@ -7,6 +7,7 @@ import Icon from "../Icon";
 import Popover from "../Popover";
 import Divider from "../Divider";
 import Box from "../Box";
+import NavItem from "./NavItem";
 
 const SideNavWrapper = styled("div")((props) =>
   css({
@@ -50,11 +51,7 @@ const SideNavItems = styled("div")((props) =>
     alignItems: "center",
     justifyContent: "flex-start",
     textAlign: "center",
-    // "&:hover, &:focus": {
-    //   "& .sideNavItemName": {
-    //     display: "block"
-    //   }
-    // },
+
     "@media screen and (max-width: 900px)": {
       height: themeGet("appScale.navBarSize")(props),
       flexDirection: "row",
@@ -67,7 +64,7 @@ const SideNavItems = styled("div")((props) =>
 
 const PanelControlTooltip = styled(Popover)((props) =>
   css({
-    alignSelf: props.showItemNames ? "flex-end" : "center",
+    alignSelf: "center",
     position: props.hideExpandedControl ? "absolute" : "relative",
     right: props.hideExpandedControl ? "r" : "initial",
     top: "r",
@@ -76,7 +73,7 @@ const PanelControlTooltip = styled(Popover)((props) =>
     borderTopColor: themeGet("colors.greyLighter")(props),
     paddingTop: props.hideExpandedControl ? "0" : "r",
     width: props.hideExpandedControl ? "auto" : "100%",
-    justifyContent: props.showItemNames ? "flex-end" : "center",
+    justifyContent: "center",
     "&:focus": {
       outline: "0"
     },
@@ -105,185 +102,6 @@ const PanelControl = styled("button")((props) =>
   })
 );
 
-const SideNavItemPopover = styled(Popover)((props) =>
-  css({
-    width: "100%",
-    marginBottom: props.bottomAligned ? "0" : "s",
-    marginTop: props.bottomAligned ? "s" : "0",
-    "&:hover,&:focus-within": {
-      "& .popoverText": {
-        opacity: props.showItemNames ? "0" : "1",
-        zIndex: props.showItemNames ? "-100" : "100",
-        visibility: props.showItemNames ? "hidden" : "visible",
-        pointerEvents: props.showItemNames ? "none" : "auto",
-        display: props.showItemNames ? "none" : "initial"
-      }
-    },
-    "&:focus-within": {
-      "& .popoverText": {
-        opacity: props.active ? "0" : "1",
-        zIndex: props.active ? "-100" : "100",
-        visibility: props.active ? "hidden" : "visible",
-        pointerEvents: props.active ? "none" : "auto",
-        display: props.active ? "none" : "initial"
-      }
-    },
-    "@media screen and (max-width: 900px)": {
-      width: "auto",
-      marginBottom: "0",
-      marginTop: "0",
-      "&:hover,&:focus-within": {
-        "& .popoverText": {
-          opacity: "0",
-          zIndex: "-100",
-          visibility: "hidden",
-          pointerEvents: "none",
-          justifyContent: "space-around",
-          display: "none"
-        }
-        // "&:hover, &:focus": {
-        //   "& .sideNavItemName": {
-        //     display: "none"
-        //   }
-        // }
-      }
-    }
-  })
-);
-
-const SideNavItem = styled("button")(
-  (props) =>
-    css({
-      fontFamily: themeGet("fonts.main")(props),
-      position: "relative",
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-      width: "100%",
-      borderRadius: themeGet("radii.2")(props),
-      transition: themeGet("transition.transitionDefault")(props),
-      bg: "transparent",
-      cursor: "pointer",
-      border: "none",
-      padding: "s",
-      // marginBottom: props.bottomAligned ? "0" : "s",
-      // marginTop: props.bottomAligned ? "s" : "0",
-      fontSize: "1.4rem",
-      fontWeight: themeGet("fontWeights.1")(props),
-      color: themeGet("colors.greyDarker")(props),
-      path: {
-        transition: themeGet("transition.transitionDefault")(props),
-        fill: themeGet("colors.greyDarker")(props)
-      },
-      "&:hover, &:focus": {
-        outline: "0",
-        color: themeGet("colors.primary")(props),
-        path: {
-          fill: themeGet("colors.primary")(props)
-        },
-        "& .sideNavItemName": {
-          color: themeGet("colors.primary")(props)
-        }
-        // "& .sideNavItemName": {
-        //   color: themeGet("colors.primary")(props)
-        // }
-      },
-      "@media screen and (max-width: 900px)": {
-        width: "auto"
-      }
-    }),
-  (props) =>
-    props.active &&
-    css({
-      bg: themeGet("colors.primary")(props),
-      path: {
-        fill: themeGet("colors.white")(props)
-      },
-      // "& .sideNavItemName": {
-      //   color: themeGet("colors.white")(props)
-      // },
-      "&:hover, &:focus": {
-        path: {
-          fill: themeGet("colors.white")(props)
-        },
-        "& .sideNavItemName": {
-          color: themeGet("colors.white")(props)
-        }
-        // "& .sideNavItemName": {
-        //   color: themeGet("colors.white")(props)
-        // }
-      }
-    })
-);
-
-const SideNavItemLink = styled("div")(
-  (props) =>
-    css({
-      "& > a": {
-        cursor: "pointer",
-        padding: "s",
-        textDecoration: "none",
-        borderRadius: themeGet("radii.2")(props),
-        width: "100%",
-        path: {
-          transition: themeGet("transition.transitionDefault")(props),
-          fill: themeGet("colors.greyDarker")(props)
-        },
-        fontFamily: themeGet("fonts.main")(props),
-        position: "relative",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        // width: "100%",
-        transition: themeGet("transition.transitionDefault")(props),
-        bg: "transparent",
-        fontSize: "1.4rem",
-        fontWeight: themeGet("fontWeights.1")(props),
-        "&:hover, &:focus": {
-          "& .sideNavItemName": {
-            color: themeGet("colors.primary")(props)
-          },
-          path: {
-            fill: themeGet("colors.primary")(props)
-          }
-        },
-        "@media screen and (max-width: 900px)": {
-          width: "auto"
-        },
-        "&:focus": {
-          outline: "0",
-          color: themeGet("colors.primary")(props),
-          path: {
-            fill: themeGet("colors.primary")(props)
-          }
-        }
-      }
-    }),
-  (props) =>
-    props.active &&
-    css({
-      "& > a": {
-        bg: themeGet("colors.primary")(props),
-        path: {
-          fill: themeGet("colors.white")(props)
-        },
-        "&:hover, &:focus": {
-          path: {
-            fill: themeGet("colors.white")(props)
-          }
-          // "& .sideNavItemName": {
-          //   color: themeGet("colors.white")(props)
-          // }
-        },
-        "&:focus": {
-          path: {
-            fill: themeGet("colors.white")(props)
-          }
-        }
-      }
-    })
-);
-
 const BottomAlignedNavItems = styled("div")(() =>
   css({
     marginTop: "auto",
@@ -297,49 +115,6 @@ const BottomAlignedNavItems = styled("div")(() =>
     }
   })
 );
-
-const BadgeNumber = styled("div")((props) =>
-  css({
-    position: "absolute",
-    height: "16px",
-    width: "16px",
-    bg: themeGet("colors.danger")(props),
-    fontSize: "1rem",
-    fontWeight: themeGet("fontWeights.2")(props),
-    color: themeGet("colors.white")(props),
-    borderRadius: "50%",
-    top: "0",
-    right: "0",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center"
-  })
-);
-const BadgeDot = styled("div")((props) =>
-  css({
-    position: "absolute",
-    height: "8px",
-    width: "8px",
-    bg: themeGet("colors.primary")(props),
-    borderRadius: "50%",
-    top: "2px",
-    right: "0"
-  })
-);
-
-// const SideNavItemName = styled("div")((props) =>
-//   css({
-//     lineHeight: "0",
-//     marginLeft: themeGet("space.r")(props),
-//     fontSize: themeGet("fontSizes.1")(props),
-//     whiteSpace: "nowrap",
-//     textDecoration: "none",
-//     color: themeGet("colors.greyDarker")(props),
-//     "&:hover, &:focus": {
-//       color: themeGet("colors.primary")(props)
-//     }
-//   })
-// );
 
 const SideNavExpanded = styled("div")((props) =>
   css({
@@ -390,85 +165,13 @@ const SideNavExpanded = styled("div")((props) =>
   })
 );
 
-const NavItem = ({
-  item,
-  Component,
-  activeItem,
-  handleItemClick,
-  // showItemNames
-  ref
-}) => {
-  // Use the ternary operator to render the appropriate component based on actionType
-  const accessibilityProps = {
-    ...(item.actionType === "component" && {
-      "aria-expanded": item.index === activeItem ? "true" : "false"
-    }),
-    "aria-label": item.name
-  };
-  return item.actionType === "link" ? (
-    <SideNavItemPopover
-      text={item.name}
-      textAlign="center"
-      width="124px"
-      active={item.index === activeItem}
-      bottomAligned={item.bottomAligned}
-      // showItemNames={showItemNames}
-    >
-      <SideNavItemLink
-        key={item.index}
-        active={item.index === activeItem}
-        bottomAligned={item.bottomAligned}
-        onClick={() => handleItemClick(item)}
-        ref={ref}
-      >
-        <Component item={item}>
-          <Icon icon={["far", item.iconName]} />
-          {/* {showItemNames && (
-            <SideNavItemName className="sideNavItemName">
-              {item.name}
-            </SideNavItemName>
-          )} */}
-        </Component>
-      </SideNavItemLink>
-    </SideNavItemPopover>
-  ) : (
-    <SideNavItemPopover
-      text={item.name}
-      textAlign="center"
-      width="124px"
-      active={item.index === activeItem}
-      bottomAligned={item.bottomAligned}
-      // showItemNames={showItemNames}
-    >
-      <SideNavItem
-        key={item.index}
-        active={item.index === activeItem}
-        onClick={() => handleItemClick(item)}
-        bottomAligned={item.bottomAligned}
-        {...accessibilityProps}
-        ref={ref}
-      >
-        {item.badgeNumber && <BadgeNumber>{item.badgeNumber}</BadgeNumber>}
-        {item.badgeDot && <BadgeDot />}
-        <Icon icon={["far", item.iconName]} />
-        {/* {showItemNames && (
-          <SideNavItemName className="sideNavItemName">
-            {item.name}
-          </SideNavItemName>
-        )} */}
-      </SideNavItem>
-    </SideNavItemPopover>
-  );
-};
-
 const SideNav = ({ items, sideNavHeight }) => {
   const [expandedItem, setExpandedItem] = useState(null);
   const [activeItem, setActiveItem] = useState(null);
 
-  // Initialize a ref for SideNavExpanded and navItem
+  // Initialize a ref for SideNavExpanded
   const expandedRef = useRef(null);
-  // const navItemRef = useRef(null);
-  const [navItemRefs, setNavItemRefs] = useState([]);
+  const navItemRefs = useRef({});
 
   const handleItemClick = (item) => {
     const { index: itemIndex, actionType, onClick: onButtonClick } = item;
@@ -476,21 +179,13 @@ const SideNav = ({ items, sideNavHeight }) => {
       setExpandedItem(null);
       onButtonClick && onButtonClick(item);
     } else {
-      onButtonClick && onButtonClick(item);
-      // expandedRef.current = null;
-
-      console.log({
-        itemIndex,
-        expandedItem,
-        navItemRefs: navItemRefs
-      });
-
-      console.log({ item });
-
-      navItemRefs[itemIndex] &&
-        navItemRefs[itemIndex].current &&
-        navItemRefs[itemIndex].current.focus();
       setExpandedItem(itemIndex === expandedItem ? null : itemIndex);
+      onButtonClick && onButtonClick(item);
+
+      itemIndex === expandedItem &&
+        navItemRefs.current &&
+        navItemRefs.current[itemIndex] &&
+        navItemRefs.current[itemIndex].focus();
     }
 
     setActiveItem(itemIndex === activeItem ? null : itemIndex);
@@ -498,15 +193,6 @@ const SideNav = ({ items, sideNavHeight }) => {
 
   // Split items into two arrays based on the bottomAligned prop
   const allItems = items.map((item, index) => ({ ...item, index }));
-
-  useEffect(() => {
-    // Add or remove refs as needed
-    setNavItemRefs((navItemRefs) =>
-      Array(allItems.length)
-        .fill()
-        .map((_, i) => navItemRefs[i] || createRef())
-    );
-  }, [allItems.length]);
 
   const topAlignedItems = allItems.filter(
     (item) => !item.bottomAligned && !item.pageSpecific
@@ -542,44 +228,13 @@ const SideNav = ({ items, sideNavHeight }) => {
     }
   }, [expandedItem]);
 
-  // Toggle nav item names on click of panel control button
-  // const [showItemNames, setShowItemNames] = useState(false);
-  // const toggleItemNames = () => {
-  //   setShowItemNames((prevState) => !prevState);
-  // };
+  const handleBlur = (item) => {
+    handleItemClick(item);
+  };
 
   return (
     <SideNavWrapper sideNavHeight={sideNavHeight}>
       <SideNavItems>
-        {/* <Flex
-          width="100%"
-          alignItems="center"
-          mb="s"
-          justifyContent={showItemNames === true ? "space-between" : "center"}
-          display={["none", "none", "none", "flex"]}
-        >
-          <PanelControlTooltip
-            width="178px"
-            textAlign="center"
-            showItemNames={showItemNames}
-            text={
-              showItemNames === true
-                ? "Hide navigation item names"
-                : "Show navigation item names"
-            }
-          >
-            <PanelControl
-              onClick={toggleItemNames}
-              aria-label="toggle navigation item names"
-            >
-              {showItemNames === true ? (
-                <Icon icon={["fas", "chevron-left"]} />
-              ) : (
-                <Icon icon={["fas", "chevron-right"]} />
-              )}
-            </PanelControl>
-          </PanelControlTooltip>
-        </Flex> */}
         {topAlignedItems.map((item) => {
           if (item.hide) {
             return null;
@@ -590,10 +245,9 @@ const SideNav = ({ items, sideNavHeight }) => {
               key={item.index}
               item={item}
               Component={Component}
-              // showItemNames={showItemNames}
               activeItem={activeItem}
               handleItemClick={handleItemClick}
-              ref={navItemRefs[item.index]}
+              navItemRefs={navItemRefs}
             />
           );
         })}
@@ -612,56 +266,15 @@ const SideNav = ({ items, sideNavHeight }) => {
                   key={item.index}
                   item={item}
                   Component={Component}
-                  // showItemNames={showItemNames}
                   activeItem={activeItem}
                   handleItemClick={handleItemClick}
-                  ref={navItemRefs[item.index]}
+                  navItemRefs={navItemRefs}
                 />
               );
             })}
           </>
         )}
-        {/* {bottomAlignedItems.length > 0 && // Render the special-container only if there are bottom-aligned items
-          (isGreaterThan900 ? (
-            <>
-              <BottomAlignedNavItems>
-                {bottomAlignedItems.map((item) => {
-                  const Component = item.component;
-                  return (
-                    <NavItem
-                      key={item.index}
-                      item={item}
-                      Component={Component}
-                      showItemNames={showItemNames}
-                      activeItem={activeItem}
-                      handleItemClick={handleItemClick}
-                    />
-                  );
-                })}
-              </BottomAlignedNavItems>
-              <PanelControlTooltip
-                width="180px"
-                textAlign="center"
-                mt="r"
-                showItemNames={showItemNames}
-                text={
-                  showItemNames === true
-                    ? "Hide navigation item names"
-                    : "Show navigation item names"
-                }
-              >
-                <PanelControl
-                  onClick={toggleItemNames}
-                  aria-label="toggle navigation item names"
-                >
-                  {showItemNames === true ? (
-                    <Icon icon={["fas", "chevron-left"]} />
-                  ) : (
-                    <Icon icon={["fas", "chevron-right"]} />
-                  )}
-                </PanelControl>
-              </PanelControlTooltip>
-            </> */}
+
         {bottomAlignedItems.length > 0 && // Render the special-container only if there are bottom-aligned items
           (isGreaterThan900 ? (
             <BottomAlignedNavItems>
@@ -675,10 +288,9 @@ const SideNav = ({ items, sideNavHeight }) => {
                     key={item.index}
                     item={item}
                     Component={Component}
-                    // showItemNames={showItemNames}
                     activeItem={activeItem}
                     handleItemClick={handleItemClick}
-                    ref={navItemRefs[item.index]}
+                    navItemRefs={navItemRefs}
                   />
                 );
               })}
@@ -694,10 +306,9 @@ const SideNav = ({ items, sideNavHeight }) => {
                   key={item.index}
                   item={item}
                   Component={Component}
-                  // showItemNames={showItemNames}
                   activeItem={activeItem}
                   handleItemClick={handleItemClick}
-                  ref={navItemRefs[item.index]}
+                  navItemRefs={navItemRefs}
                 />
               );
             })
@@ -724,6 +335,7 @@ const SideNav = ({ items, sideNavHeight }) => {
             <PanelControl
               onClick={() => handleItemClick(allItems[expandedItem])}
               aria-label="toggle panel"
+              onBlur={() => handleBlur(allItems[expandedItem])}
             >
               {isGreaterThan900 === true ? (
                 <Icon icon={["fas", "chevron-left"]} />
@@ -736,15 +348,6 @@ const SideNav = ({ items, sideNavHeight }) => {
       ) : null}
     </SideNavWrapper>
   );
-};
-
-NavItem.propTypes = {
-  item: PropTypes.object,
-  Component: PropTypes.elementType,
-  activeItem: PropTypes.string,
-  handleItemClick: PropTypes.func,
-  // showItemNames: PropTypes.bool
-  ref: PropTypes.object
 };
 
 SideNav.propTypes = {
@@ -763,8 +366,7 @@ SideNav.propTypes = {
       onClick: PropTypes.func
     })
   ).isRequired,
-  LinkComponent: PropTypes.elementType // React Router Link component
+  LinkComponent: PropTypes.elementType
 };
 
-/** @component */
 export default SideNav;
